move save stuff to save.cpp
